LayoutTests/webrtc/video-mute-expected.txt

Video should be running, go to black and running.
Following, should be a snapshot of the video, a black frame and a snapshot of the video.
  

PASS Setting video exchange 
PASS Track is enabled, video should not be black 
PASS Track is disabled, video should be black 
PASS Track is enabled, video should not be black 
